requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装


Django==4.2.16
djangorestframework==3.14.0
django-redis==5.4.0
celery==5.2.7
redis==4.6.0
django-filter==23.3
psycopg2-binary==2.9.6
